Add files via upload
Correct Files committed, last commit was the wrong set. Added a Menu, and 2 end scenes depending on if you Seven out or Crap out. Added sound when dice are rolled. Menu has instructions on how to play, and end scenes rest to the menu. Added comments in code, and changed variable names to be more understandable.